refactor(web): move api-status into components/marketing
ApiStatus is imported only by the (marketing) landing, so it belongs in the fork-excluded
components/marketing bucket (the earlier audit mis-classified it as core; the landing move made
it clear). No .gitpickignore change needed: components/marketing is already excluded.
